Add explicit disambiguation for Item/ItemGroup
Add explicit disambiguation for Item/ItemGroup by attempting to access the getRelativeNameFromGroup API that is made available starting in 1.515.  If this method is not available, we fall back on what is there today, which cannot properly disambiguate between the Item overload and the ItemGroup overload when "context" is both (as is the case with some unreleased plugins).

I have tested this both with Jenkins core at 1.509.3 and at head (1.569).