make fillDialog work with URLs that start with /
fillDialog was up until now prepending the URL with a slash regardless
of the actual URL argument. If it got called with a URL starting with /,
the resulting URL would start with double slashes and have its first
directory interpreted as a domain by the browser, causing a "server not
found" error on the interface.